Refactored code. Moved styles to 2 different files and moved text to separate files. Imported all into RenderLandingForGroomers file.
