+++ src/components/pages/LandingForGroomers/Text.js
@@ -1,0 +1,30 @@
+const PageMessage = `Join our network of professional groomers. We will display your
+    business to clients in your area. They will be able to schedule
+    appointments with you. You will be able to check the client's pets
+    to make sure they have the appropriate vaccines and verifications
+    before going to your establishment. Once the service is finished the
+    client will be able to provide feedback and rate the service they
+    received. Do a good job and your rating will increase making your
+    business visible to more clients in your area. This is an
+    extraordinary opportunity to make your business widespread and
+    available to more clients.`;
+
+const ReviewOne = `John Doe: “If you're looking for a place to expand your grooming
+    business this is it. I started out in my garage by myself. Today I
+    own a large building with several employees. I even have a groomer
+    bus that allows me to go to the client’s house. It was all thanks to
+    Express Groomer. ”`;
+
+const ReviewTwo = `Jane Miller: “I recently started using Express Groomer. I wanted
+    more people to know about my business. Within a few days I started
+    seeing a lot more people coming into my establishment and many more
+    scheduling appointments through Express Groomer. Thank you Express
+    Groomer!”`;
+
+const ReviewThree = `James Smith: “Thanks to Express Groomer I am able to check my
+    client’s pets for verifications and vaccines. If a specific pet does
+    not meet the necessary requirements. I can cancel the appointment
+    and request the client to take the pet to a vet before coming into
+    my business.”`;
+
+export { PageMessage, ReviewOne, ReviewTwo, ReviewThree };
